Make BNW run on Factorio 0.17
Some fixes to avoid the first crashes I encountered with Factorio 0.17
The quickbar no longer exists as an entity that can hold items

There's still much to improve, especially integrating this with the new
ghost functionality instead of creating blueprints for indivitual items.

Also it seems you can now duplicate power poles (I managed it with the
small ones).

Gameplay-wise it might be good to include a turret in the starting
inventory since the locals are much more aggressive, though I'm not sure
about it.
I managed to barely survive the first attack, but lost an electric miner
in the process.  It seems to me that it could be manageable as-is, but
you *really* have to research gun turrets first thing.